requirements.txt
huggingface_hub==0.7.0
matplotlib==3.3.4
numpy==1.19.2
opencv_python==4.5.4.58
Pillow==9.1.1
pydensecrf==1.0rc3
scikit_image==0.16.2
scipy==1.5.4
skimage==0.0
timm==0.4.5
torch==1.7.1
torchvision==0.8.2
visdom==0.1.8.9
